feat(pipeline): update battery level to integer and round values for consistency

+++ src/main/java/cloud/chlora/pipeline/telemetry/domain/model/Telemetry.java
@@ -12,7 +12,7 @@ public record Telemetry(
         float soilMoisture,
         float temperature,
         float humidity,
-        float batteryLevel,
+        int batteryLevel,
         Instant receivedAt,
         boolean isValid
 ) {}
